refactor: move data generate functions to data module
