+++ tests/conftest.py
@@ -12,7 +12,7 @@ from expertsystem.amplitude.helicity import HelicityModel
 from expertsystem.amplitude.kinematics import ReactionInfo
 from expertsystem.particle import ParticleCollection
 
-from tensorwaves.data.generate import generate_data, generate_phsp
+from tensorwaves.data import generate_data, generate_phsp
 from tensorwaves.data.helicity import HelicityKinematicsConverter
 from tensorwaves.data.phasespace import TFUniformRealNumberGenerator
 from tensorwaves.estimator import SympyUnbinnedNLL
