+++ tests/data/test_generate.py
@@ -8,7 +8,7 @@ from expertsystem.amplitude.data import EventCollection
 from expertsystem.amplitude.kinematics import ReactionInfo
 from expertsystem.particle import ParticleCollection
 
-from tensorwaves.data.generate import generate_phsp
+from tensorwaves.data import generate_phsp
 from tensorwaves.data.phasespace import TFUniformRealNumberGenerator
